+++ frappe/Help/help.js
@@ -1,5 +1,4 @@
 import { ref } from 'vue'
 
 export const showHelpModal = ref(false)
-export const showHelpCenter = ref(false)
 export const minimize = ref(false)
