+++ frappe/HelpCenter/HelpCenter.vue
@@ -62,9 +62,9 @@
   </div>
 </template>
 <script setup>
-import Button from '../../components/Button/Button.vue'
-import FeatherIcon from '../../components/FeatherIcon.vue'
-import TextInput from '../../components/TextInput.vue'
+import Button from '../../src/components/Button/Button.vue'
+import FeatherIcon from '../../src/components/FeatherIcon.vue'
+import TextInput from '../../src/components/TextInput.vue'
 import { ref, computed, onMounted } from 'vue'
 
 const props = defineProps({
