+++ frappe/HelpCenter/helpCenter.js
@@ -1,0 +1,3 @@
+import { ref } from 'vue'
+
+export const showHelpCenter = ref(false)
