+++ frappe/Icons/LightningIcon.vue
