+++ frappe/Icons/MaximizeIcon.vue
