+++ frappe/Icons/MinimizeIcon.vue
