+++ frappe/Icons/StepsIcon.vue
