+++ frappe/Onboarding/GettingStartedBanner.vue
@@ -52,11 +52,12 @@
   </Button>
 </template>
 <script setup>
-import StepsIcon from './Icons/StepsIcon.vue'
-import Button from '../../components/Button/Button.vue'
-import FeatherIcon from '../../components/FeatherIcon.vue'
-import { useOnboarding } from '../composables/onboarding'
-import { showHelpCenter, showHelpModal, minimize } from '../composables/help'
+import StepsIcon from '../Icons/StepsIcon.vue'
+import Button from '../../src/components/Button/Button.vue'
+import FeatherIcon from '../../src/components/FeatherIcon.vue'
+import { useOnboarding } from './onboarding'
+import { showHelpCenter } from '../HelpCenter/helpCenter'
+import { showHelpModal, minimize } from '../Help/help'
 
 const props = defineProps({
   isSidebarCollapsed: {
