+++ frappe/index.js
@@ -1,0 +1,19 @@
+// help components
+export { default as HelpModal } from './Help/HelpModal.vue'
+
+// onboarding components
+export { default as GettingStartedBanner } from './Onboarding/GettingStartedBanner.vue'
+export { default as OnboardingSteps } from './Onboarding/OnboardingSteps.vue'
+
+// help center components
+export { default as HelpCenter } from './HelpCenter/HelpCenter.vue'
+
+// billing components
+export { default as TrialBanner } from './Billing/TrialBanner.vue'
+
+// composables
+export { useOnboarding } from './Onboarding/onboarding.js'
+
+// utils
+export { showHelpModal, minimize } from './Help/help.js'
+export { showHelpCenter } from './HelpCenter/helpCenter.js'
