+++ tailwind.config.js
@@ -1,7 +1,12 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   presets: [require('./src/tailwind/preset')],
-  content: ['./index.html', './App.vue', './src/**/*.{vue,js,ts,jsx,tsx}'],
+  content: [
+    './index.html',
+    './App.vue',
+    './src/**/*.{vue,js,ts,jsx,tsx}',
+    './frappe/**/*.{vue,js,ts,jsx,tsx}',
+  ],
   theme: {
     extend: {},
   },
